refactor: rework project structure
