+++ .github/workflows/build.yml
@@ -17,7 +17,7 @@ jobs:
       - uses: actions/checkout@v3
 
       - name: Verify Formatting
-        run: clang-format -i src/**/*.cpp src/**/*.hpp --dry-run -Werror
+        run: clang-format -i apps/**/*.cpp include/**/*.hpp --dry-run -Werror
 
       - name: Build IceFlow Applications
         run: |
