+++ .gitignore
@@ -85,6 +85,7 @@ CMakeFiles
 CMakeCache.txt
 cmake_install.cmake
 Makefile
+install_manifest.txt
 
 # VS Code files
 .vscode
