+++ .gitlab-ci.yml
@@ -8,7 +8,7 @@ Verify formatting with clang-format:
   script:
     - apt-get update
     - apt-get -y install clang-format
-    - clang-format -i src/**/*.cpp src/**/*.hpp --dry-run -Werror
+    - clang-format -i apps/**/*.cpp include/**/*.hpp --dry-run -Werror
 
 build:
   stage: build
