+++ apps/Aggregate/aggregate.cpp
@@ -16,12 +16,13 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-#include "core/consumer-tlv.hpp"
-#include "core/measurements.hpp"
-#include "yaml-cpp/yaml.h"
-
 #include <csignal>
 #include <thread>
+
+#include "yaml-cpp/yaml.h"
+
+#include "iceflow/consumer-tlv.hpp"
+#include "iceflow/measurements.hpp"
 
 // ###### MEASUREMENT ######
