+++ apps/PeopleCounter/.gitignore
