+++ apps/PeopleCounter/config.yaml
