+++ apps/PeopleCounter/peoplecounter.cpp
@@ -16,14 +16,15 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-#include "core/consumer-tlv.hpp"
-#include "core/measurements.hpp"
-#include "core/producer-tlv.hpp"
+#include <csignal>
+#include <thread>
+
 #include "opencv2/objdetect.hpp"
 #include "yaml-cpp/yaml.h"
 
-#include <csignal>
-#include <thread>
+#include "iceflow/consumer-tlv.hpp"
+#include "iceflow/measurements.hpp"
+#include "iceflow/producer-tlv.hpp"
 
 // ###### MEASUREMENT ######
