+++ cmake/iceflow-config.cmake
@@ -1,0 +1,10 @@
+include(CMakeFindDependencyMacro)
+
+find_dependency(Boost REQUIRED)
+find_dependency(PkgConfig)
+pkg_check_modules(libndn-cxx REQUIRED)
+pkg_check_modules(PSYNC REQUIRED PSync)
+
+get_filename_component(SELF_DIR ${CMAKE_CURRENT_LIST_DIR} PATH)
+
+include(${SELF_DIR}/cmake/IceFlow.cmake)
